refactor(#374): extract shared extract_contract_id() into provisioner module
Move duplicated extract_contract_id() from proxmox.rs, docker.rs, and
digitalocean.rs into provisioner/mod.rs. Consolidate duplicate tests
into a single test in the shared module.

Net change: -40 lines of duplication, +24 lines of shared code.
